+++ vagrant-playground/server2/_deploy_loki_grafana_prometheus.sh
@@ -5,7 +5,8 @@ PROJECT_FOLDER="/srv/loki_prometheus_grafana/"
 
 mkdir -p ${PROJECT_FOLDER}
 
-cp ${PROJECT_FOLDER}/loki-config.yaml ${PROJECT_FOLDER}/loki-config.yaml.bak
+# Deploy loki-config.yaml
+[ -f ${PROJECT_FOLDER}/loki-config.yaml ] && cp ${PROJECT_FOLDER}/loki-config.yaml ${PROJECT_FOLDER}/loki-config.yaml.bak
 cat <<'EOF' > "${PROJECT_FOLDER}/loki-config.yaml"
 auth_enabled: false
 
@@ -63,6 +64,34 @@ EOF
 
 loki_config_modified=$(if [ "$(md5sum ${PROJECT_FOLDER}/loki-config.yaml | awk '{ print $1 }')" != "$(md5sum ${PROJECT_FOLDER}/loki-config.yaml.bak | awk '{ print $1 }')" ]; then echo true; else echo false; fi)
 
+# Deploy prometheus-config.yaml
+[ -f ${PROJECT_FOLDER}/prometheus-config.yaml ] && cp ${PROJECT_FOLDER}/prometheus-config.yaml ${PROJECT_FOLDER}/prometheus-config.yaml.bak
+cat <<'EOF' > "${PROJECT_FOLDER}/prometheus-config.yaml"
+global:
+  scrape_interval: 30s
+
+scrape_configs:
+  - job_name: 'prometheus'
+    static_configs:
+      - targets: 
+        - 'localhost:9090'
+
+  - job_name: 'node_exporter'
+    static_configs:
+      - targets:
+        - 'server1.vagrant.test:9100'
+        - 'server2.vagrant.test:9100'
+
+  - job_name: 'cadvisor'
+    static_configs:
+      - targets: 
+        - 'server1.vagrant.test:8080'
+        - 'server2.vagrant.test:8080'
+EOF
+
+prometheus_config_modified=$(if [ "$(md5sum ${PROJECT_FOLDER}/prometheus-config.yaml | awk '{ print $1 }')" != "$(md5sum ${PROJECT_FOLDER}/prometheus-config.yaml.bak | awk '{ print $1 }')" ]; then echo true; else echo false; fi)
+
+
 cat <<'EOF' > "${PROJECT_FOLDER}docker-compose.yaml"
 services:
   loki:
@@ -70,8 +99,7 @@ services:
     image: grafana/loki:3.2.0
     restart: unless-stopped
     user: root
-    ports:
-      - "3100:3100"
+    network_mode: "host"
     command: "-config.file=/etc/loki/local-config.yaml"
     volumes:
       - ./loki-config.yaml:/etc/loki/local-config.yaml
@@ -87,17 +115,37 @@ services:
       - GF_SECURITY_ADMIN_PASSWORD=password
       - GF_USERS_DEFAULT_THEME=light
       - GF_USERS_ALLOW_SIGN_UP=false
-    ports:
-      - "3000:3000"
+    network_mode: "host"
     volumes:
       - /var/lib/grafana:/var/lib/grafana
+
+  prometheus:
+    container_name: prometheus
+    image: quay.io/prometheus/prometheus:v2.55.0
+    restart: unless-stopped
+    user: root
+    volumes:
+      - ./prometheus-config.yaml:/etc/prometheus/prometheus.yaml
+      - /var/lib/prometheus/:/prometheus
+    network_mode: "host"
+    command:
+      - "--config.file=/etc/prometheus/prometheus.yaml"
+      - "--storage.tsdb.path=/prometheus"
 EOF
+
+ufw allow 3100 # Loki
+ufw allow 3000 # Grafana
+ufw allow 9090 # Prometheus
 
 cd ${PROJECT_FOLDER}
 docker compose pull
 
 if [ "$loki_config_modified" == "true" ]; then
     docker compose stop loki
+fi
+
+if [ "$prometheus_config_modified" == "true" ]; then
+    docker compose stop prometheus
 fi
 
 docker compose up -d
